pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "npc-talk-demo-flat"
version = "0.1.0"
description = "Minimal NPC group chat demo (Agentscope, Kimi) with world tools"
requires-python = ">=3.11"
readme = "README.md"
authors = [{ name = "npc-talk maintainers" }]
dependencies = [
    # Agentscope from Git (prefer conda env per README, but keep here for pip workflows)
    "agentscope @ git+https://github.com/agentscope-ai/agentscope.git",
    # Web framework
    "Flask>=2.3.0",
    "Flask-CORS>=4.0.0",
    # Authentication
    "PyJWT>=2.8.0",
    # Configuration and data handling
    "PyYAML>=6.0",
    # File monitoring
    "watchdog>=3.0.0",
    # Performance and reporting
    "matplotlib>=3.7.0",
    "pandas>=2.0.0",
    "psutil>=5.9.0",
    # Additional utilities
    "aiohttp>=3.8.0",
    "python-dateutil>=2.8.0",
]

[project.optional-dependencies]
dev = [
    "ruff>=0.6",
    "mypy>=1.9",
    "pytest>=8.0",
    "pytest-asyncio>=0.21.0",
    "pytest-cov>=4.1.0",
    "pytest-mock>=3.10.0",
    "types-setuptools",
]

## No console scripts; run with `python src/main.py`

[tool.ruff]
line-length = 100
target-version = "py311"
exclude = ["src/world/tools.py"]

[tool.ruff.lint]
select = ["E", "F", "W", "I" ]
ignore = ["E402"]

[tool.mypy]
python_version = "3.11"
strict = false
warn_unused_ignores = true
warn_redundant_casts = true
warn_unreachable = true
exclude = ["src/world/tools.py"]
